Create pull_request_template.md
This PR template surfaces the trivial patch exemption immediately so that the wall-of-text CLA is hopefully less intrusive.

It also adds a notification that contributions may receive public recognition in release notes. With this change, Legal has signed off on us providing recognition and on IAC continuing to provide recognition.

Language is totally open for wordsmithing.