(maint) Issue Template Update
I received a report in puppetlabs/puppetlabs-release that using an EOL
operating system (Ubuntu 18) as an example had mildly bad optics, so I
found the source of it here.

My proposal is to not force ourselves to face this problem every
now-and-then by removing the specific example but still attempting to
have helpful text.

Additionally, I don't feel the phrase 'clear and consise' is clear and
concise, so I trimmed it along with bits of other verbiage in the
hopes of making the templates themselves concise.